Separate TokenIndex + related classes.
